requirements.txt
google-generativeai>=0.3.0
edge-tts>=6.1.0
gtts>=2.5.0
pyttsx3>=2.90
playwright>=1.40.0
apscheduler>=3.10.0
ffmpeg-python>=0.2.0
pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
